+++ tests/test_routes.py
@@ -300,111 +300,6 @@ class TestYourResourceService(TestCase):
         self.assertEqual(get_order["status"], "SHIPPED")
         self.assertNotIn("extra_field", get_order)
 
-    def test_cancel_order(self):
-        """change status of order to CANCELLED"""
-        item = ItemFactory()
-        order = item.order
-        order.status = "CREATED"
-        order.create()
-
-        response = self.client.put(f"orders/{order.id}/cancel")
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-    def test_cancel_bad_order(self):
-        """cancel order that does not exist"""
-        order = OrderFactory()
-        order.status = "CREATED"
-        order.create()
-
-        response = self.client.put("orders/0/cancel")
-        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
-
-    def test_update_order_status(self):
-        """update the status of an order"""
-        order = OrderFactory()
-        order.status = "PENDING"
-        order.create()
-
-        # create item and assert
-        item = ItemFactory()
-
-        response = self.client.post(
-            f"orders/{order.id}/items",
-            json=item.serialize(),
-            content_type="application/json",
-        )
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-
-        response = self.client.put(f"orders/{order.id}/update")
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(order.status.value, "CREATED")
-
-    def test_update_bad_order(self):
-        """Try to update order that does not exist"""
-        order = OrderFactory()
-        order.status = "CREATED"
-        order.create()
-
-        # create item and assert
-        item = ItemFactory()
-
-        response = self.client.post(
-            f"orders/{order.id}/items",
-            json=item.serialize(),
-            content_type="application/json",
-        )
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-
-        response = self.client.put("orders/0/update")
-        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
-
-    def test_update_empty_order(self):
-        """Update order with empty item list"""
-        order = OrderFactory()
-        order.status = "CREATED"
-        order.create()
-
-        response = self.client.put(f"orders/{order.id}/update")
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-
-    def test_update_completed_order(self):
-        """Try to update order that is completed"""
-        order = OrderFactory()
-        order.status = "COMPLETED"
-        order.create()
-
-        # create item and assert
-        item = ItemFactory()
-
-        response = self.client.post(
-            f"orders/{order.id}/items",
-            json=item.serialize(),
-            content_type="application/json",
-        )
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-
-        response = self.client.put(f"orders/{order.id}/update")
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-
-    def test_update_cancelled_order(self):
-        """Try to update order that is cancelled"""
-        order = OrderFactory()
-        order.status = "CANCELLED"
-        order.create()
-
-        # create item and assert
-        item = ItemFactory()
-
-        response = self.client.post(
-            f"orders/{order.id}/items",
-            json=item.serialize(),
-            content_type="application/json",
-        )
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-
-        response = self.client.put(f"orders/{order.id}/update")
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-
     # TEST CASES FOR CREATING ITEMS ########################
 
     def test_create_item(self):
